Fixing __main__ issues
