+++ puremagic/__init__.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-from .__main__ import *
-from .__main__ import __version__, __author__
+from puremagic.main import __version__, __author__
+from puremagic.main import *
+
